Ultrasonic Sensor Files
Ultrasonic sensor datasheet uploaded. C code written to measure the distance of object using ultrasonic sensor, turning on an LED when the object is within 4cm of the sensor. This will be used to detect a block in the loading bay. We can edit the triggerAction function to control the Dobot rather than an LED, once we work out the exact Dobot functions.