C Code for RBG Sensor
Contains a fully functional C code for the RGB Sensor. The code sets the colour filters on the sensor, reads the individual RGB values and dictates the colour based on the RGB combination values (outputted as DetectedColour). 50 samples of DetectedColour are taken, and the average is outputted as DetectedColourAverage. This is the value that will be used to command the Dobot to move a block to its relevant colour area. The code only does this the once, then ends.